requirements.txt
streamlit
pandas
plotly
google-cloud-bigquery 
streamlit-option-menu
google-genai  
db-dtypes
